+++ src/class/UIObjects/TextButton.cpp
@@ -48,22 +48,22 @@ void TextButton::draw()
 
     if (press)
     {
-        DrawText(text.c_str(), rect.x, rect.y, fontSize, pressColor);
+        DrawText(text.c_str(), static_cast<int>(rect.x), static_cast<int>(rect.y), fontSize, pressColor);
     }
     else if (hover)
     {
-        DrawText(text.c_str(), rect.x, rect.y, fontSize, hoverColor);
+        DrawText(text.c_str(), static_cast<int>(rect.x), static_cast<int>(rect.y), fontSize, hoverColor);
     }
     else
     {
-        DrawText(text.c_str(), rect.x, rect.y, fontSize, normalColor);
+        DrawText(text.c_str(), static_cast<int>(rect.x), static_cast<int>(rect.y), fontSize, normalColor);
     }
     
 
 #if _DEBUG
-    DrawLine(position.x, 0, position.x, GetScreenHeight(), WHITE);
-    DrawLine(0, position.y, GetScreenWidth(), position.y, WHITE);
-    DrawCircle(position.x, position.y, 5.0f, WHITE);
+    DrawLine(static_cast<int>(position.x), 0, static_cast<int>(position.x), GetScreenHeight(), WHITE);
+    DrawLine(0, static_cast<int>(position.y), GetScreenWidth(), static_cast<int>(position.y), WHITE);
+    DrawCircle(static_cast<int>(position.x), static_cast<int>(position.y), 5.0f, WHITE);
 #endif
 }
